Added more

+++ Lib/inc/i2c.h
@@ -25,7 +25,8 @@ typedef enum I2C_Speed_Typedef__
 void I2C_Init(I2C_TypeDef* i2c, I2C_Speed speed);
 void I2C_Reset(I2C_TypeDef* i2c);
 int I2C_IsBusy(I2C_TypeDef* i2c);
-int I2C_Transmit(I2C_TypeDef* i2c, uint16_t addr, uint8_t* pData, uint8_t size);
+int I2C_Transmit(I2C_TypeDef* i2c, uint8_t addr, uint8_t* pData, uint8_t size);
+
